backend/requirements.txt
Django==5.1.8
djangorestframework==3.15.2
psycopg[binary]==3.2.6
python-dotenv==1.0.1
